els-ASAP auto commit file/tag libmwaw-0.3.5-1.el7.src.rpm

+++ src/src/tools/file/rsrc.cpp
@@ -31,7 +31,6 @@
 #include <sstream>
 #include <string>
 
-#include <libmwaw_internal.hxx>
 #include "file_internal.h"
 #include "input.h"
 #include "rsrc.h"
@@ -42,8 +41,8 @@ std::string RSRC::Version::makePretty(std::string const &orig)
 {
   size_t j;
   std::string res(orig);
-  for (; (j = res.find( 0xd )) != std::string::npos;)
-    res.replace( j, 1, " -- " );
+  for (; (j = res.find(0xd)) != std::string::npos;)
+    res.replace(j, 1, " -- ");
   return res;
 }
 
@@ -147,7 +146,7 @@ std::vector<RSRC::MapEntry> RSRC::getMapEntries(std::string type)
     if (lEntry.m_pos >= 0 && lEntry.m_pos+4 < eof)
       res.push_back(lEntry);
     else {
-      MWAW_DEBUG_MSG(("RSRC::getMapEntries: find bad pos %lx\n", lEntry.m_pos));
+      MWAW_DEBUG_MSG(("RSRC::getMapEntries: find bad pos %lx\n", (long unsigned int) lEntry.m_pos));
     }
     m_input.seek(pos+12, InputStream::SK_SET);
   }
